+++ components/donation/DonationPage.tsx
@@ -6,8 +6,8 @@ import ContainerWithShadow from "components/ContainerWithShadow";
 import { Nonprofit, DropdownProps } from "utils/types";
 
 import DonationPageLayout from "./DonationPageLayout";
-import DonationPageFormHeader from "./DonationPageFormHeader";
-import DonationPageFormBody from "./DonationPageFormBody";
+import DonationPageNonprofitBanner from "./DonationPageNonprofitBanner";
+import DonationPageMainContent from "./DonationPageMainContent";
 
 const useStyles = makeStyles({
   container: {
@@ -26,12 +26,12 @@ const DonationPage: React.FC<Props & DropdownProps> = ({
 }) => {
   const { container } = useStyles();
 
-  // Setting key will tell React to re-mount the DonationPageFormBody component when the selected Nonprofit in the dropdown changes
+  // Setting key will tell React to re-mount the DonationPageMainContent component when the selected Nonprofit in the dropdown changes
   return (
     <DonationPageLayout {...dropdownProps}>
       <ContainerWithShadow className={container}>
-        <DonationPageFormHeader headline={nonprofit.headline} />
-        <DonationPageFormBody
+        <DonationPageNonprofitBanner headline={nonprofit.headline} />
+        <DonationPageMainContent
           description={nonprofit.about}
           selectedNonprofitId={dropdownProps.selectedValue}
           key={dropdownProps.selectedValue}
